Fix corners in tabbed/stacked root containers
Titlebars of containers in a tabbed/stacking layout are adjusted so that
only the outer corners are rounded. Previously this was done by checking
the current layout of the parent container. However, if the layout of
the workspace is set to tabbed/stacking then the root containers in the
workspace will be tabbed/stacked without having a parent container, and
as such their parent is NULL and their corners weren't being adjusted.

With this commit, if the parent is NULL then the layout of the current
workspace is checked to adjust the corners of tabbed/stacked children.