chore: remove unnecessary items from compose file & update build
The `version` field is long deprecated.

The `image` field is superfluous at best, misleading at worst: When the
`--build` flag is not specified, which happens especially to those not
familiar with Docker and Docker Compose, Compose will try to fetch the
image from DockerHub, but this will fail since there is no such image on
DockerHub.

Lastly, `working_dir` is already specified in the `Dockerfile`.

We will see in an upcoming commit how setting the environment variables
as `ARG`s in _one place_ is better than setting them in two places.

Signed-off-by: Georg Lauterbach <44545919+georglauterbach@users.noreply.github.com>