fix: resolve ShellCheck complaints in build.sh
Using `/bin/sh` is great for portability, but seeing that it was
introduces for the purpose of using it with the Docker build, it incures
unnecessary friction. `/bin/bash` is a way better choice, expecially
seeing that Debian Trixie is used as the base image.

The variable globbing was fixed with proper quoting, and variable
expansion was fixed with proper braces.

I updated the HEREDOC too. The original authors probably did not know
about the `"EOF"` vs `EOF` syntax. With `"EOF"`, variables are not
expanded in the HEREDOC.

Signed-off-by: Georg Lauterbach <44545919+georglauterbach@users.noreply.github.com>